moved tests to test folder to match in consistency with other apps
