+++ apps/portal/test/SigninFlow.test.js
@@ -1,12 +1,13 @@
-import App from '../App.js';
-import {fireEvent, appRender, within, waitFor} from '../utils/test-utils';
-import {site as FixtureSite} from '../utils/test-fixtures';
-import setupGhostApi from '../utils/api.js';
+import App from '../src/App.js';
+import {fireEvent, appRender, within, waitFor} from '../src/utils/test-utils';
+import {site as FixtureSite} from '../src/utils/test-fixtures';
+import setupGhostApi from '../src/utils/api.js';
 
 const OTC_LABEL_REGEX = /Code/i;
 
 const setup = async ({site, member = null, labs = {}}) => {
     const ghostApi = setupGhostApi({siteUrl: 'https://example.com'});
+
     ghostApi.init = vi.fn(() => {
         return Promise.resolve({
             site,
