+++ apps/portal/test/portal-links.test.js
@@ -1,11 +1,12 @@
-import App from '../App';
-import {site as FixtureSite, member as FixtureMember} from '../utils/test-fixtures';
-import {appRender, within} from '../utils/test-utils';
-import setupGhostApi from '../utils/api';
+import App from '../src/App';
+import {site as FixtureSite, member as FixtureMember} from '../src/utils/test-fixtures';
+import {appRender, within} from '../src/utils/test-utils';
+import setupGhostApi from '../src/utils/api';
 import {fireEvent} from '@testing-library/react';
 
 const setup = async ({site, member = null, showPopup = true}) => {
     const ghostApi = setupGhostApi({siteUrl: 'https://example.com'});
+
     ghostApi.init = vi.fn(() => {
         return Promise.resolve({
             site,
